Break up tests per method, add custom shared + views tests
