+++ test/integration/default/shared-tests.js
@@ -1,0 +1,381 @@
+let { dirname, join } = require('path')
+let {
+  existsSync,
+  lstatSync,
+  mkdirSync,
+  readFileSync,
+  renameSync,
+  writeFileSync
+} = require('fs')
+let cp = require('cpr')
+let test = require('tape')
+let {
+  resetAndCopyShared,
+  resetAndCopySharedCustom,
+  checkFolderCreation,
+  arcFileArtifacts,
+  staticArtifacts,
+  sharedArtifacts,
+  sharedArtifactsDisabled,
+  getViewsArtifacts,
+  viewsArtifacts,
+  viewsArtifactsDisabled,
+} = require('../_shared')
+let hydrate = require('../../..')
+process.env.CI = true // Suppresses tape issues with progress indicator
+
+test(`[Shared file copying (default paths)] shared() never uses symlinks by default`, t => {
+  t.plan(2)
+  resetAndCopyShared(t, function () {
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        let path = 'src/http/get-index/node_modules/@architect/shared'
+        let file = path + '/shared.md'
+        let stat = lstatSync(path).isSymbolicLink()
+        t.notOk(stat, 'shared directory was copied, and is not a symlink')
+        t.equal(readFileSync(file).toString(), 'It me!', 'Copied file is readable')
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (custom paths)] shared() never uses symlinks by default`, t => {
+  t.plan(2)
+  resetAndCopySharedCustom(t, function () {
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        let path = 'src/http/get-index/node_modules/@architect/shared'
+        let file = path + '/shared.md'
+        let stat = lstatSync(path).isSymbolicLink()
+        t.notOk(stat, 'shared directory was copied, and is not a symlink')
+        t.equal(readFileSync(file).toString(), 'It me!', 'Copied file is readable')
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (default paths)] shared() copies src/shared and src/views (unless disabled or folder not found; @views not specified)`, t => {
+  t.plan(
+    sharedArtifacts.length +
+    getViewsArtifacts.length +
+    sharedArtifactsDisabled.length +
+    viewsArtifactsDisabled.length + 1
+  )
+  resetAndCopyShared(t, function () {
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        // Check to see if files that are supposed to be there are actually there
+        sharedArtifacts.forEach(path => {
+          t.ok(existsSync(path), `Found shared file in ${path}`)
+        })
+        sharedArtifactsDisabled.forEach(path => {
+          t.notOk(existsSync(path), `Did not find shared file in function with @arc shared false ${path}`)
+        })
+        getViewsArtifacts.forEach(path => {
+          if (path.includes('get-')) {
+            t.ok(existsSync(path), `Found views file in GET function ${path}`)
+          }
+          else {
+            t.notOk(existsSync(path), `Did not find views file in non-GET function ${path}`)
+          }
+        })
+        viewsArtifactsDisabled.forEach(path => {
+          t.notOk(existsSync(path), `Did not find views file in function with @arc views false ${path}`)
+        })
+        checkFolderCreation(t)
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (custom paths)] shared() copies src/shared and src/views (unless disabled or folder not found; @views not specified)`, t => {
+  t.plan(
+    sharedArtifacts.length +
+    getViewsArtifacts.length +
+    sharedArtifactsDisabled.length +
+    viewsArtifactsDisabled.length + 1
+  )
+  resetAndCopySharedCustom(t, function () {
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        // Check to see if files that are supposed to be there are actually there
+        sharedArtifacts.forEach(path => {
+          t.ok(existsSync(path), `Found shared file in ${path}`)
+        })
+        sharedArtifactsDisabled.forEach(path => {
+          t.notOk(existsSync(path), `Did not find shared file in function with @arc shared false ${path}`)
+        })
+        getViewsArtifacts.forEach(path => {
+          if (path.includes('get-')) {
+            t.ok(existsSync(path), `Found views file in GET function ${path}`)
+          }
+          else {
+            t.notOk(existsSync(path), `Did not find views file in non-GET function ${path}`)
+          }
+        })
+        viewsArtifactsDisabled.forEach(path => {
+          t.notOk(existsSync(path), `Did not find views file in function with @arc views false ${path}`)
+        })
+        checkFolderCreation(t)
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (default paths)] shared() src/views to only @views (unless disabled or folder not found)`, t => {
+  t.plan(viewsArtifacts.length + viewsArtifactsDisabled.length + 1)
+  resetAndCopyShared(t, function () {
+    cp(join('src', 'app.arc-views'), join('.', 'app.arc'), { overwrite: true }, function (err) {
+      if (err) t.fail(err)
+      else {
+        hydrate.shared({}, function (err) {
+          if (err) t.fail(err)
+          else {
+            // Check to see if files that are supposed to be there are actually there
+            viewsArtifacts.forEach(path => {
+              if (path.includes('get-index') || path.includes('post-up-tents')) {
+                t.ok(existsSync(path), `Found views file in GET function ${path}`)
+              }
+              else {
+                t.notOk(existsSync(path), `Did not find views file in non-GET function ${path}`)
+              }
+            })
+            viewsArtifactsDisabled.forEach(path => {
+              t.notOk(existsSync(path), `Did not find views file in function with @arc views false ${path}`)
+            })
+            checkFolderCreation(t)
+          }
+        })
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (custom paths)] shared() src/views to only @views (unless disabled or folder not found)`, t => {
+  t.plan(viewsArtifacts.length + viewsArtifactsDisabled.length + 1)
+  resetAndCopySharedCustom(t, function () {
+    cp(join('_shared-custom', 'app.arc-views'), join('.', 'app.arc'), { overwrite: true }, function (err) {
+      if (err) t.fail(err)
+      else {
+        hydrate.shared({}, function (err) {
+          if (err) t.fail(err)
+          else {
+            // Check to see if files that are supposed to be there are actually there
+            viewsArtifacts.forEach(path => {
+              if (path.includes('get-index') || path.includes('post-up-tents')) {
+                t.ok(existsSync(path), `Found views file in GET function ${path}`)
+              }
+              else {
+                t.notOk(existsSync(path), `Did not find views file in non-GET function ${path}`)
+              }
+            })
+            viewsArtifactsDisabled.forEach(path => {
+              t.notOk(existsSync(path), `Did not find views file in function with @arc views false ${path}`)
+            })
+            checkFolderCreation(t)
+          }
+        })
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (default paths)] shared() copies app.arc file and static.json (Arc <5)`, t => {
+  t.plan(arcFileArtifacts.length + staticArtifacts.length + 1)
+  process.env.DEPRECATED = true
+  resetAndCopyShared(t, function () {
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        delete process.env.DEPRECATED
+        // Check to see if files that are supposed to be there are actually there
+        arcFileArtifacts.forEach(path => {
+          t.ok(existsSync(path), `Found app.arc file in ${path}`)
+        })
+        staticArtifacts.forEach(path => {
+          t.ok(existsSync(path), `Found static.json file in ${path}`)
+        })
+        checkFolderCreation(t)
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (custom paths)] shared() copies app.arc file and static.json (Arc <5)`, t => {
+  t.plan(arcFileArtifacts.length + staticArtifacts.length + 1)
+  process.env.DEPRECATED = true
+  resetAndCopySharedCustom(t, function () {
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        delete process.env.DEPRECATED
+        // Check to see if files that are supposed to be there are actually there
+        arcFileArtifacts.forEach(path => {
+          t.ok(existsSync(path), `Found app.arc file in ${path}`)
+        })
+        staticArtifacts.forEach(path => {
+          t.ok(existsSync(path), `Found static.json file in ${path}`)
+        })
+        checkFolderCreation(t)
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (default paths)] shared() copies static.json but not app.arc (Arc v6+)`, t => {
+  t.plan(arcFileArtifacts.length + staticArtifacts.length + 1)
+  resetAndCopyShared(t, function () {
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        // Check to see if files that are supposed to be there are actually there
+        arcFileArtifacts.forEach(path => {
+          t.notOk(existsSync(path), `Did not find app.arc file in ${path}`)
+        })
+        staticArtifacts.forEach(path => {
+          t.ok(existsSync(path), `Found static.json file in ${path}`)
+        })
+        checkFolderCreation(t)
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (custom paths)] shared() copies static.json but not app.arc (Arc v6+)`, t => {
+  t.plan(arcFileArtifacts.length + staticArtifacts.length + 1)
+  resetAndCopySharedCustom(t, function () {
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        // Check to see if files that are supposed to be there are actually there
+        arcFileArtifacts.forEach(path => {
+          t.notOk(existsSync(path), `Did not find app.arc file in ${path}`)
+        })
+        staticArtifacts.forEach(path => {
+          t.ok(existsSync(path), `Found static.json file in ${path}`)
+        })
+        checkFolderCreation(t)
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (default paths)] shared() copies static.json with @static folder configured`, t => {
+  t.plan(staticArtifacts.length + 3)
+  resetAndCopyShared(t, function () {
+    // Rewrite app.arc to include @static folder directive
+    let arcFile = join(process.cwd(), 'app.arc')
+    let arc = readFileSync(arcFile).toString()
+    arc += '@static\nfolder foo'
+    writeFileSync(arcFile, arc)
+    t.pass(`Added '@static folder foo' to app.arc`)
+    // Move public/ to foo/
+    renameSync(join(process.cwd(), 'public'), join(process.cwd(), 'foo'))
+    t.ok(existsSync(join(process.cwd(), 'foo', 'static.json')), 'public/static.json moved into foo/static.json')
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        // Check to see if files that are supposed to be there are actually there
+        staticArtifacts.forEach(path => {
+          t.ok(existsSync(path), `Found static.json file in ${path}`)
+        })
+        checkFolderCreation(t)
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (custom paths)] shared() copies static.json with @static folder configured`, t => {
+  t.plan(staticArtifacts.length + 3)
+  resetAndCopySharedCustom(t, function () {
+    // Rewrite app.arc to include @static folder directive
+    let arcFile = join(process.cwd(), 'app.arc')
+    let arc = readFileSync(arcFile).toString()
+    arc += '@static\nfolder foo'
+    writeFileSync(arcFile, arc)
+    t.pass(`Added '@static folder foo' to app.arc`)
+    // Move public/ to foo/
+    renameSync(join(process.cwd(), 'public'), join(process.cwd(), 'foo'))
+    t.ok(existsSync(join(process.cwd(), 'foo', 'static.json')), 'public/static.json moved into foo/static.json')
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        // Check to see if files that are supposed to be there are actually there
+        staticArtifacts.forEach(path => {
+          t.ok(existsSync(path), `Found static.json file in ${path}`)
+        })
+        checkFolderCreation(t)
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (default paths)] shared() should remove files in functions that do not exist in src/shared and src/views`, t => {
+  t.plan(sharedArtifacts.length + getViewsArtifacts.length + 1)
+  resetAndCopyShared(t, function () {
+    let sharedStragglers = sharedArtifacts.map((p) => {
+      let dir = dirname(p)
+      mkdirSync(dir, { recursive: true })
+      let file = join(dir, 'straggler.json')
+      writeFileSync(file, '{surprise:true}')
+      return file
+    })
+    let viewsStragglers = getViewsArtifacts.map((p) => {
+      let dir = dirname(p)
+      mkdirSync(dir, { recursive: true })
+      let file = join(dir, 'straggler.json')
+      writeFileSync(file, '{surprise:true}')
+      return file
+    })
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        // Check to see if files that are supposed to be there are actually there
+        sharedStragglers.forEach(path => {
+          t.notOk(existsSync(path), `shared straggler file removed from ${path}`)
+        })
+        viewsStragglers.forEach(path => {
+          t.notOk(existsSync(path), `views straggler file removed from ${path}`)
+        })
+        checkFolderCreation(t)
+      }
+    })
+  })
+})
+
+test(`[Shared file copying (custom paths)] shared() should remove files in functions that do not exist in src/shared and src/views`, t => {
+  t.plan(sharedArtifacts.length + getViewsArtifacts.length + 1)
+  resetAndCopySharedCustom(t, function () {
+    let sharedStragglers = sharedArtifacts.map((p) => {
+      let dir = dirname(p)
+      mkdirSync(dir, { recursive: true })
+      let file = join(dir, 'straggler.json')
+      writeFileSync(file, '{surprise:true}')
+      return file
+    })
+    let viewsStragglers = getViewsArtifacts.map((p) => {
+      let dir = dirname(p)
+      mkdirSync(dir, { recursive: true })
+      let file = join(dir, 'straggler.json')
+      writeFileSync(file, '{surprise:true}')
+      return file
+    })
+    hydrate.shared({}, function (err) {
+      if (err) t.fail(err)
+      else {
+        // Check to see if files that are supposed to be there are actually there
+        sharedStragglers.forEach(path => {
+          t.notOk(existsSync(path), `shared straggler file removed from ${path}`)
+        })
+        viewsStragglers.forEach(path => {
+          t.notOk(existsSync(path), `views straggler file removed from ${path}`)
+        })
+        checkFolderCreation(t)
+      }
+    })
+  })
+})
