+++ test/mock/_shared/app.arc-views
@@ -1,4 +1,4 @@
-# _shared-custom + specified views
+# _shared
 
 @app
 hydrate
@@ -38,8 +38,6 @@ rivers
 @streams
 rivers
 
-@shared
-src custom-shared-folder
-
 @views
-src custom-views-folder
+get /
+post /up/tents
